+++ Transportation/Entities/ProductInput.cs
@@ -14,7 +14,7 @@ namespace Transportation
         public long InputOrderID { get; set; }
         public virtual InputOrder InputOrder { get; set; }
         public virtual Product Product { get; set; }
-        public long Quantity { get; set; }
+        public double Quantity { get; set; }
         public long InputPrice { get; set; }
         public long SalePrice { get; set; }
 
@@ -45,7 +45,7 @@ namespace Transportation
             ID = json.Value<long>("id");
             ProductID = json.Value<long>("productId");
             InputOrderID = json.Value<long>("inputOrderID");
-            Quantity = json.Value<long>("quantity");
+            Quantity = json.Value<double>("quantity");
             InputPrice = json.Value<long>("inputPrice");
             SalePrice = json.Value<long>("salePrice");
         }
